Split up assessment forms
