+++ barriers/views/assessments/overview.py
@@ -1,0 +1,7 @@
+from django.views.generic import TemplateView
+
+from ..mixins import BarrierMixin
+
+
+class AssessmentOverview(BarrierMixin, TemplateView):
+    template_name = "barriers/assessments/overview.html"
